requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
coverage==7.0.5
Django==4.1.5
django-betterforms==2.0.0
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-model-utils==4.3.1
django-widget-tweaks==1.4.12
ez-setup==0.9
Faker==16.4.0
fontawesomefree==6.2.1
idna==3.4
jsonfield==3.1.0
newsapi-python==0.2.6
numpy==1.24.2
Pillow==9.4.0
pydantic==1.10.4
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
swapper==1.3.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
